requirements.txt
flask==3.1.2
pdfplumber==0.11.8
pymupdf==1.26.7
pytesseract==0.3.13
Pillow==12.0.0
spacy==3.8.11
